Fixed broken tests

+++ Bullet-Inferno-test/src/se/dat255/bulletinferno/model/entity/EnemyImplTest.java
@@ -122,7 +122,7 @@ public class EnemyImplTest {
 
 		// Set another enemy as source and check that friendly fire is't enabled
 		projectile.setSource(enemy2);
-		int preCollisionHealth = enemy.getHealth();
+		float preCollisionHealth = enemy.getHealth();
 
 		// Simulate colliding with a projectile
 		enemy.preCollided(projectile);
